fixed bug where deleteItem could be null

+++ EventTranslator.js
@@ -108,12 +108,13 @@ function eventTranslator(changeEvent) {
 
             // iterate over all remaining deletions
             for(let deleteItem of checksums.deletions) {
-                console.log(`[Delete]: ${JSON.stringify(deleteItem)}`);
-                // if we didn't find a match above, this must be a real deletion
-                translated.push({
-                    Data: {collectionName, delete: deleteItem.data},
-                    PartitionKey: deleteItem.checksum
-                });
+                if(deleteItem) {
+                    // if we didn't find a match above, this must be a real deletion
+                    translated.push({
+                        Data: {collectionName, delete: deleteItem.data},
+                        PartitionKey: deleteItem.checksum
+                    });
+                }
             }
 
         } catch (err) {
